+++ test/test_evolve.jl
@@ -38,7 +38,7 @@
     end
     @testset "Empty environment and sequence" begin
         simulation = mr.Simulation(mr.Sequence(TR=2.8))
-        snaps = mr.trajectory(zeros(3), simulation, 0:0.5:2.8)
+        snaps = mr.readout(zeros(3), simulation, 0:0.5:2.8, return_snapshot=true)
         time = 0.
         for snap in snaps
             @test snap.time == time
@@ -50,7 +50,7 @@
         @test length(snaps) == 6
 
         simulation = mr.Simulation(mr.Sequence(TR=2.8))
-        snaps= mr.trajectory([mr.Spin(), mr.Spin()], simulation, 0:0.5:2.8)
+        snaps= mr.readout([mr.Spin(), mr.Spin()], simulation, 0:0.5:2.8, return_snapshot=true)
         time = 0.
         for snap in snaps
             @test snap.time == time
@@ -63,7 +63,7 @@
     end
     @testset "Gradient echo sequence" begin
         simulation = mr.Simulation(mr.Sequence(components=[mr.InstantRFPulse(flip_angle=90)], TR=2.8))
-        snaps = mr.trajectory(zeros(3), simulation, 0:0.5:2.8)
+        snaps = mr.readout(zeros(3), simulation, 0:0.5:2.8)
         @test mr.orientation(snaps[1]) ≈ SA[0., 0., 1.]
         for snap in snaps[2:end]
             @test mr.orientation(snap) ≈ SA[0., 1., 0.]
@@ -71,15 +71,16 @@
         @test length(snaps) == 6
     end
     @testset "Ensure data is stored at requested time" begin
-        simulation = mr.Simulation([mr.Sequence(TR=2.8)])
-        snaps = mr.trajectory(mr.Spin(), simulation, 0:0.5:2.8)
-        @test length(snaps) == 6
+        simulation = mr.Simulation(mr.Sequence(TR=2.8))
 
         snaps = mr.evolve(mr.Spin(), simulation, 2.3)
         @test mr.get_time(snaps) == 2.3
 
         snaps = mr.evolve(snaps, simulation)
         @test mr.get_time(snaps) == 2.8
+
+        snaps = mr.evolve(snaps, simulation)
+        @test mr.get_time(snaps) == 5.6
 
         snaps = mr.evolve(mr.Spin(), simulation)
         @test mr.get_time(snaps) == 2.8
@@ -100,8 +101,9 @@
         sequence = mr.Sequence(components=[mr.InstantRFPulse(flip_angle=90)], TR=2.)
         sphere = mr.spheres(radius=1.)
         Random.seed!(12)
-        diff = mr.Simulation([mr.Sequence(TR=20.)], diffusivity=2., geometry=sphere)
-        snaps = mr.trajectory([mr.Spin(), mr.Spin()], diff, 0:0.5:sequence.TR)
+        diff = mr.Simulation(mr.Sequence(TR=20.), diffusivity=2., geometry=sphere)
+        snaps = mr.readout([mr.Spin(), mr.Spin()], diff, 0:0.5:sequence.TR, return_snapshot=true)
+        @test size(snaps) == (5, )
         for snap in snaps
             @test length(snap.spins) == 2
             for spin in snap.spins
@@ -118,7 +120,8 @@
         ]
         all_snaps = mr.Simulation(sequences, diffusivity=1., R2=1.)
 
-        readouts = [r[1] for r in mr.readout(mr.Spin(), all_snaps)]
+        readouts = mr.readout(mr.Spin(), all_snaps)
+        @assert size(readouts) == (3,)
 
         # check relaxation
         @test mr.transverse(readouts[1]) ≈ 0. atol=1e-12